Move source tree conftest.py contents to top level
Having a conftest.py in the module apparently caused the issues
with import location error, which was not a false positive after all.
