+++ tox.ini
@@ -11,10 +11,6 @@ deps=
 extras=
     hdbscan
     hdfs
-setenv=
-    # Somehow the conftest.py file in src/libertem leads to a collision
-    # See https://github.com/pytest-dev/pytest/issues/2042
-    PY_IGNORE_IMPORTMISMATCH=1
 
 [testenv:flake8]
 changedir={toxinidir}
